added support for function input

+++ graph drawer/src/CoordinateManager.h
@@ -3,20 +3,20 @@
 class CoordinateManager
 {
 public:
-	static long long int GetCenterX();
-	static long long int GetCenterY();
-	static long long int GetScaleLevel();
-	static long long int ScreenToWorldX(int screenX);
-	static long long int ScreenToWorldY(int screenY);
-	static int WorldToScreenX(long long int worldX);
-	static int WorldToScreenY(long long int worldY);
+	static double GetCenterX();
+	static double GetCenterY();
+	static double GetScaleLevel();
+	static double ScreenToWorldX(int screenX);
+	static double ScreenToWorldY(int screenY);
+	static int WorldToScreenX(double worldX);
+	static int WorldToScreenY(double worldY);
 	static float ScreenNormalizedX(int screenX);
 	static float ScreenNormalizedY(int screenY);
 	static void Update();
 private:
 	CoordinateManager();
 
-	static long long int centerX, centerY;
-	static long long int scaleLevel;
-	static long long int panOldX, panOldY;
+	static double centerX, centerY;
+	static double scaleLevel;
+	static double panOldX, panOldY;
 };
